windows 现在可以成功编译了， 但是编译结果还没有进行测试， 以及dx11 还没有应用

+++ sight_widgets.h
@@ -42,12 +42,12 @@ namespace sight {
     };
 
     class SightKey : public SightEmptyKey {
-        ushort code{0};
+        int code{0};
 
     public:
         SightKey() = default;
         SightKey(SightKey const& key) = default;
-        SightKey(ushort code);
+        SightKey(int code);
         ~SightKey() = default;
 
         bool isKeyReleased() const override;
